backend/requirements.txt
aniso8601==7.0.0
argon2==0.1.10
argon2-cffi==20.1.0
asgiref==3.2.10
bcrypt==3.1.7
certifi==2020.6.20
cffi==1.14.0
chardet==3.0.4
defusedxml==0.6.0
dj-database-url==0.5.0
Django==3.0.1
django-allauth==0.41.0
django-cors-headers==3.2.1
django-environ==0.4.5
django-filter==2.3.0
django-graphql-jwt==0.3.0
graphene==2.1.8
graphene-django==2.9.0
graphql-core==2.3.2
graphql-relay==2.0.1
gunicorn==20.0.4
idna==2.10
oauthlib==3.1.0
promise==2.3
psycopg2-binary==2.8.4
pycparser==2.20
PyJWT==1.7.1
python-http-client==3.2.7
python3-openid==3.2.0
pytz==2020.1
requests==2.24.0
requests-oauthlib==1.3.0
Rx==1.6.1
sendgrid==6.4.3
singledispatch==3.4.0.3
six==1.15.0
sqlparse==0.3.1
starkbank-ecdsa==1.0.0
urllib3==1.25.9
whitenoise==5.0.1